Net Interface Updated

+++ src/mem/ruby/network/garnet2.0/NetworkInterface.hh
@@ -93,8 +93,8 @@ class NetworkInterface : public ClockedObject, public Consumer,
     int sim_init(AstraSim::AstraMemoryAPI* MEM);
     AstraSim::timespec_t sim_get_time();
     void sim_schedule(AstraSim::timespec_t delta, void (*fun_ptr)(void *fun_arg), void *fun_arg);
-    int sim_send(void *buffer, int count, int type, int dst, int tag, AstraSim::sim_request *request, void (*msg_handler)(void *fun_arg), void* fun_arg);
-    int sim_recv(void *buffer, int count, int type, int src, int tag, AstraSim::sim_request *request, void (*msg_handler)(void *fun_arg), void* fun_arg);
+    int sim_send(void *buffer, uint64_t count, int type, int dst, int tag, AstraSim::sim_request *request, void (*msg_handler)(void *fun_arg), void* fun_arg);
+    int sim_recv(void *buffer, uint64_t count, int type, int src, int tag, AstraSim::sim_request *request, void (*msg_handler)(void *fun_arg), void* fun_arg);
     AstraNetworkAPI::BackendType get_backend_type(){return AstraNetworkAPI::BackendType::Garnet;};
     
     class Recv_Req{
